fixbug: txcount
Former-commit-id: 66d800945a964bb1eb4d1dff4805f43b76670dbe
Former-commit-id: 9dc1bf2b6eb3c901b24534bef89fd9b4ed4a0d28

+++ types/types.go
@@ -354,7 +354,7 @@ func (block *Block) GetHeader() *Header {
 	if head.Height >= ForkBlockHash {
 		head.Difficulty = block.Difficulty
 		head.StateHash = block.StateHash
-		head.TxCount = len(block.Txs)
+		head.TxCount = int64(len(block.Txs))
 	}
 	return head
 }
